[nanda] update middleware login for admin and customer
