+++ routes/auth.js
@@ -4,5 +4,4 @@ const authController = require('../controllers/authController');
 router.post('/login', authController.login);
 router.post('/register', authController.register);
 
-
 module.exports = router
