+++ routes/index.js
@@ -4,7 +4,7 @@ const productRouter = require('./product');
 const categoryRouter = require('./category');
 const bannerRouter = require('./banner');
 
-const isLogin = require('../middlewares/isLogin');
+const isLogin = require('../middlewares/isLoginAdmin');
 
 router.get('/', (req, res) => {
   res.status(200).json({
@@ -14,7 +14,6 @@ router.get('/', (req, res) => {
 })
 
 router.use(authRouter);
-router.use(isLogin)
 router.use('/products', productRouter);
 router.use('/categories', categoryRouter);
 router.use('/banners', bannerRouter);
